requirements.txt
fastapi==0.111.0
passlib==1.7.4
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
SQLAlchemy==2.0.29
bcrypt==4.0.1
uvicorn==0.22.0
pydantic==1.10.11
python-jose==3.3.0
pyjwt==2.8.0
